recommendation-engine/api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.25.1,<3.0.0
python-dotenv>=0.19.0
tensorflow>=2.8.0
tensorflow-recommenders>=0.7.2
pandas>=1.3.0
numpy>=1.19.5
scikit-learn>=1.0.0
pymongo>=4.0.0
aiohttp>=3.6.2
urllib3>=1.26,<3
cachetools>=4.0.0
